release: 0.5.0 — device coupling, per-unit scaling, warm start, seeded placement
A minor rather than a patch release. It adds battery converter reactive power,
warm starting, seeded device placement and three add_OPF options, and changes
five defaults — each reversible with one argument. CHANGELOG.md carries the full
list; the headline items:

Fixed

- pandapower 3.5 moved create_continuous_bus_index and clear_result_tables out
  of its top-level namespace, which broke every single-period constructor and
  left the base-case power flow in net.res_* on both result mappers. solve()
  also wrapped the termination check and the result mapping in one
  except AttributeError, so the mapper failure was logged into a disabled logger
  and discarded — which is what hid it.
- Multi-period solve(to_net=True) logged a mapping it never performed, so
  callers read base-case values as the optimisation result.
- The multi-period battery applied its one-way efficiency to a single signed
  power variable, so a charge/discharge cycle lost nothing at any efficiency,
  and nothing constrained the final state of charge.
- Multi-period sgen parameters were in MW while demand was per-unit, so the
  nodal balance mixed scales on any network with sn_mva != 1. A pinned model
  missed pandapower's own power flow by 6.3e-2 p.u. at sn_mva = 10.
- Battery, Heatpump and PV attached their variables but never reached the
  Kirchhoff constraints, so their dispatch could not affect voltages, loading or
  the objective. PV's active power also had the opposite sign to its own Q
  characteristics, and every PV Q(U) constraint was silently skipped.
- A cold-started multi-period AC OPF could report infeasible on a model that is
  demonstrably feasible; solve() now seeds a consistent operating point from a
  power flow at each step.
- Device placement drew from the unseeded global NumPy generator, so no study
  could be re-run. It changed conclusions, not just bus numbers.
- The AC model could not be built at all on a network with shunts: the
  conductance and susceptance parameters had opposite index sets.
- Filling net.res_* mutated temporaries, which pandas turns into a silent no-op
  under copy-on-write — in one case replacing a curtailed sgen dispatch with its
  uncurtailed profile.
- __version__ was hard-coded and three releases stale.
- The README, docs and architecture diagram advertised an EV module that does
  not exist, and named a heat-pump class whose real name differs in case.

Added

- Battery converter reactive power: BAT_Q bounded by the apparent-power circle,
  an optional power-factor floor, and the grid-code Q(P)/Q(U) areas, with Q(P)
  keyed on the discharging leg per VDE-AR-N 4105/4110/4120.
- warm_start_from_pf(), seed/rng on every placing device, thermal_limit,
  angle_limits and free_slack_vm on the multi-period models, map_to_net(t), and
  terminal_soc on the battery.

Changed

- solve(warm_start=True), Battery(terminal_soc="cyclic"), PV's pPV as a
  generator-convention injection, seeded placements, and unsupported add_OPF
  options raising instead of being discarded.
- environment.yaml moves to pandapower 3.5.4, bringing scipy, tqdm and geojson
  with it, so CI exercises the version a fresh install resolves to. pandapower
  and pandas both gained upper bounds so a major release cannot change results
  through a fresh install.

74 new regression tests; suite at 342 here. The LPAC formulation and the
research code are not part of this repository, so their share of the work is not
included.

+++ src/potpourri/technologies/generator.py
@@ -154,15 +154,22 @@ class Generator_multi_period(Flexibility_multi_period):
         max_p = np.full(len(self.generation_data), 1e9) / self.baseMVA
         min_p = np.full(len(self.generation_data), -1e9) / self.baseMVA
 
+        # astype(float) before fillna: these pandapower columns can carry
+        # object dtype when they hold NaN, and filling an object column
+        # downcasts, which pandas deprecates and will change. Converting first
+        # keeps the result a plain float array either way.
         for element, (f, t) in self.net._gen_order.items():
             if "max_p_mw" in self.net[element]:
                 max_p[f:t] = (
-                    self.net[element].max_p_mw.fillna(1e9).values
+                    self.net[element].max_p_mw.astype(float).fillna(1e9).values
                     / self.baseMVA
                 )
             if "min_p_mw" in self.net[element]:
                 min_p[f:t] = (
-                    self.net[element].min_p_mw.fillna(-1e9).values
+                    self.net[element]
+                    .min_p_mw.astype(float)
+                    .fillna(-1e9)
+                    .values
                     / self.baseMVA
                 )
 
@@ -191,11 +198,13 @@ class Generator_multi_period(Flexibility_multi_period):
         for element, (f, t) in self.net._gen_order.items():
             if "max_q_mvar" in self.net[element]:
                 max_q[f:t] = (
-                    self.net[element].max_q_mvar.fillna(1e9) / self.baseMVA
+                    self.net[element].max_q_mvar.astype(float).fillna(1e9)
+                    / self.baseMVA
                 )
             if "min_q_mvar" in self.net[element]:
                 min_q[f:t] = (
-                    self.net[element].min_q_mvar.fillna(-1e9) / self.baseMVA
+                    self.net[element].min_q_mvar.astype(float).fillna(-1e9)
+                    / self.baseMVA
                 )
 
         self.generation_data["max_q"] = max_q
